requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
opencv-python==4.8.1.78
numpy==1.26.2
